requirements.txt
# Core dependencies
numpy>=1.24.0
scipy>=1.10.0
matplotlib>=3.6.0

# Configuration and validation
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyyaml>=6.0

# CLI and tooling
typer>=0.9.0
rich>=13.0.0  # For better CLI output

# Visualization
plotly>=5.14.0
pandas>=2.0.0  # For data manipulation

# Math and geometry
transforms3d>=0.4.1  # For rotation conversions
sympy>=1.12  # For symbolic math (useful for Jacobians)

# Testing
pytest>=7.3.0
pytest-cov>=4.0.0

# Development tools
black>=23.0.0  # Code formatting
isort>=5.12.0  # Import sorting
mypy>=1.3.0  # Type checking

# Optional but useful
tqdm>=4.65.0  # Progress bars
joblib>=1.3.0  # Parallel processing
numba>=0.57.0  # JIT compilation for performance